add --dryRun flag to flytectl subcommands (#163)

+++ flytectl/cmd/delete/matchable_attribute_util.go
@@ -9,19 +9,27 @@ import (
 )
 
 func deleteMatchableAttr(ctx context.Context, project, domain, workflowName string,
-	deleter ext.AdminDeleterExtInterface, rsType admin.MatchableResource) error {
+	deleter ext.AdminDeleterExtInterface, rsType admin.MatchableResource, dryRun bool) error {
 	if len(workflowName) > 0 {
 		// Delete the workflow attribute from the admin. If the attribute doesn't exist , admin deesn't return an error and same behavior is followed here
-		err := deleter.DeleteWorkflowAttributes(ctx, project, domain, workflowName, rsType)
-		if err != nil {
-			return err
+		if dryRun {
+			logger.Infof(ctx, "skipping DeleteWorkflowAttributes request (dryRun)")
+		} else {
+			err := deleter.DeleteWorkflowAttributes(ctx, project, domain, workflowName, rsType)
+			if err != nil {
+				return err
+			}
 		}
 		logger.Debugf(ctx, "Deleted matchable resources from %v project and domain %v and workflow %v", project, domain, workflowName)
 	} else {
 		// Delete the project domain attribute from the admin. If the attribute doesn't exist , admin deesn't return an error and same behavior is followed here
-		err := deleter.DeleteProjectDomainAttributes(ctx, project, domain, rsType)
-		if err != nil {
-			return err
+		if dryRun {
+			logger.Infof(ctx, "skipping DeleteProjectDomainAttributes request (dryRun)")
+		} else {
+			err := deleter.DeleteProjectDomainAttributes(ctx, project, domain, rsType)
+			if err != nil {
+				return err
+			}
 		}
 		logger.Debugf(ctx, "Deleted matchable resources from %v project and domain %v", project, domain)
 	}
